pyproject.toml
[project]
name = "zheera"
version = "1.0.0"
description = "ZHEERA — Kurdish Intelligent Telegram Bot"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "python-telegram-bot[webhooks]>=20.7",
    "uvicorn>=0.29.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "anyio>=4.0.0",
    "google-genai>=1.0.0",
]
